Changed htp to htb for better position
Given that figures and tables follow right after a piece of text,
this patch changes the position to using htb instead of htp

htp will actually try and go and create a page dedicated for floats
instead of trying for the float to be placed at the bottom.
Therefore, htb is a better choice.